tests/emerge: Add async_loop and binhost fixtures
Add async_loop and binhost fixtures.
This is a refactor: some setup code is moved from conftest.py to
test_simple.py.
